- Change: drop support for < 8.2 - Change: moved to enums, promoted properties - Added: logger for more socket info - Added: slave_uuid support - Change: config no longer static - Chore: typos in README/code - Chore: replace/remove old urls from code - Chore: changed variables to underscore - Added: support caching_sha2_password - Change: BinLogServerInfo static calls removed also added method getServerInfo to MySQLReplicationFactory

+++ tests/Integration/BasicTest.php
@@ -1,4 +1,5 @@
 <?php
+
 /** @noinspection PhpPossiblePolymorphicInvocationInspection */
 
 /** @noinspection PhpUnhandledExceptionInspection */
@@ -7,7 +8,6 @@ declare(strict_types=1);
 
 namespace MySQLReplication\Tests\Integration;
 
-use MySQLReplication\BinLog\BinLogServerInfo;
 use MySQLReplication\Definitions\ConstEventType;
 use MySQLReplication\Event\DTO\DeleteRowsDTO;
 use MySQLReplication\Event\DTO\FormatDescriptionEventDTO;
@@ -18,15 +18,13 @@ use MySQLReplication\Event\DTO\UpdateRowsDTO;
 use MySQLReplication\Event\DTO\WriteRowsDTO;
 use MySQLReplication\Event\DTO\XidDTO;
 use MySQLReplication\MySQLReplicationFactory;
+use RuntimeException;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
-class BasicTest extends BaseTest
+class BasicTest extends BaseCase
 {
-    /**
-     * @test
-     */
-    public function shouldGetDeleteEvent(): void
+    public function testShouldGetDeleteEvent(): void
     {
         $this->createAndInsertValue(
             'CREATE TABLE test (id INT NOT NULL AUTO_INCREMENT, data VARCHAR (50) NOT NULL, PRIMARY KEY (id))',
@@ -42,14 +40,11 @@ class BasicTest extends BaseTest
         /** @var DeleteRowsDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(DeleteRowsDTO::class, $event);
-        self::assertEquals(1, $event->getValues()[0]['id']);
-        self::assertEquals('Hello World', $event->getValues()[0]['data']);
+        self::assertEquals(1, $event->values[0]['id']);
+        self::assertEquals('Hello World', $event->values[0]['data']);
     }
 
-    /**
-     * @test
-     */
-    public function shouldGetUpdateEvent(): void
+    public function testShouldGetUpdateEvent(): void
     {
         $this->createAndInsertValue(
             'CREATE TABLE test (id INT NOT NULL AUTO_INCREMENT, data VARCHAR (50) NOT NULL, PRIMARY KEY (id))',
@@ -65,16 +60,13 @@ class BasicTest extends BaseTest
         /** @var UpdateRowsDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(UpdateRowsDTO::class, $event);
-        self::assertEquals(1, $event->getValues()[0]['before']['id']);
-        self::assertEquals('Hello', $event->getValues()[0]['before']['data']);
-        self::assertEquals(2, $event->getValues()[0]['after']['id']);
-        self::assertEquals('World', $event->getValues()[0]['after']['data']);
+        self::assertEquals(1, $event->values[0]['before']['id']);
+        self::assertEquals('Hello', $event->values[0]['before']['data']);
+        self::assertEquals(2, $event->values[0]['after']['id']);
+        self::assertEquals('World', $event->values[0]['after']['data']);
     }
 
-    /**
-     * @test
-     */
-    public function shouldGetWriteEventDropTable(): void
+    public function testShouldGetWriteEventDropTable(): void
     {
         $this->connection->executeStatement($createExpected = 'CREATE TABLE `test` (id INTEGER(11))');
         $this->connection->executeStatement('INSERT INTO `test` VALUES (1)');
@@ -83,12 +75,12 @@ class BasicTest extends BaseTest
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertEquals($createExpected, $event->getQuery());
+        self::assertEquals($createExpected, $event->query);
 
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertEquals('BEGIN', $event->getQuery());
+        self::assertEquals('BEGIN', $event->query);
 
         /** @var TableMapDTO $event */
         self::assertInstanceOf(TableMapDTO::class, $this->getEvent());
@@ -96,21 +88,18 @@ class BasicTest extends BaseTest
         /** @var WriteRowsDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(WriteRowsDTO::class, $event);
-        self::assertEquals([], $event->getValues());
-        self::assertEquals(0, $event->getChangedRows());
+        self::assertEquals([], $event->values);
+        self::assertEquals(0, $event->changedRows);
 
         self::assertInstanceOf(XidDTO::class, $this->getEvent());
 
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertStringContainsString($dropExpected, $event->getQuery());
+        self::assertStringContainsString($dropExpected, $event->query);
     }
 
-    /**
-     * @test
-     */
-    public function shouldGetQueryEventCreateTable(): void
+    public function testShouldGetQueryEventCreateTable(): void
     {
         $this->connection->executeStatement(
             $createExpected = 'CREATE TABLE test (id INT NOT NULL AUTO_INCREMENT, data VARCHAR (50) NOT NULL, PRIMARY KEY (id))'
@@ -119,18 +108,15 @@ class BasicTest extends BaseTest
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertEquals($createExpected, $event->getQuery());
+        self::assertEquals($createExpected, $event->query);
     }
 
-    /**
-     * @test
-     */
-    public function shouldDropColumn(): void
+    public function testShouldDropColumn(): void
     {
         $this->disconnect();
 
         $this->configBuilder->withEventsOnly(
-            [ConstEventType::WRITE_ROWS_EVENT_V1, ConstEventType::WRITE_ROWS_EVENT_V2]
+            [ConstEventType::WRITE_ROWS_EVENT_V1->value, ConstEventType::WRITE_ROWS_EVENT_V2->value]
         );
 
         $this->connect();
@@ -143,39 +129,40 @@ class BasicTest extends BaseTest
         /** @var WriteRowsDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(WriteRowsDTO::class, $event);
-        self::assertEquals(['id' => 1, 'DROPPED_COLUMN_1' => null], $event->getValues()[0]);
+        self::assertEquals([
+            'id' => 1,
+            'DROPPED_COLUMN_1' => null,
+        ], $event->values[0]);
 
         $event = $this->getEvent();
         self::assertInstanceOf(WriteRowsDTO::class, $event);
-        self::assertEquals(['id' => 2], $event->getValues()[0]);
+        self::assertEquals([
+            'id' => 2,
+        ], $event->values[0]);
     }
 
-    /**
-     * @test
-     */
-    public function shouldFilterEvents(): void
+    public function testShouldFilterEvents(): void
     {
         $this->disconnect();
 
-        $this->configBuilder->withEventsOnly([ConstEventType::QUERY_EVENT]);
+        $this->configBuilder->withEventsOnly([ConstEventType::QUERY_EVENT->value]);
 
         $this->connect();
 
         self::assertInstanceOf(QueryDTO::class, $this->getEvent());
         self::assertInstanceOf(QueryDTO::class, $this->getEvent());
 
-        $this->connection->executeStatement($createTableExpected = 'CREATE TABLE test (id INTEGER(11), data VARCHAR(50))');
+        $this->connection->executeStatement(
+            $createTableExpected = 'CREATE TABLE test (id INTEGER(11), data VARCHAR(50))'
+        );
 
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertEquals($createTableExpected, $event->getQuery());
+        self::assertEquals($createTableExpected, $event->query);
     }
 
-    /**
-     * @test
-     */
-    public function shouldFilterTables(): void
+    public function testShouldFilterTables(): void
     {
         $expectedTable = 'test_2';
         $expectedValue = 'foobar';
@@ -184,7 +171,7 @@ class BasicTest extends BaseTest
 
         $this->configBuilder
             ->withEventsOnly(
-                [ConstEventType::WRITE_ROWS_EVENT_V1, ConstEventType::WRITE_ROWS_EVENT_V2]
+                [ConstEventType::WRITE_ROWS_EVENT_V1->value, ConstEventType::WRITE_ROWS_EVENT_V2->value]
             )->withTablesOnly([$expectedTable]);
 
         $this->connect();
@@ -205,20 +192,15 @@ class BasicTest extends BaseTest
 
         $event = $this->getEvent();
         self::assertInstanceOf(WriteRowsDTO::class, $event);
-        self::assertEquals($expectedTable, $event->getTableMap()->getTable());
-        self::assertEquals($expectedValue, $event->getValues()[0]['data']);
+        self::assertEquals($expectedTable, $event->tableMap->table);
+        self::assertEquals($expectedValue, $event->values[0]['data']);
     }
 
-    /**
-     * @test
-     */
-    public function shouldTruncateTable(): void
+    public function testShouldTruncateTable(): void
     {
         $this->disconnect();
 
-        $this->configBuilder->withEventsOnly(
-            [ConstEventType::QUERY_EVENT]
-        );
+        $this->configBuilder->withEventsOnly([ConstEventType::QUERY_EVENT->value]);
 
         $this->connect();
 
@@ -230,19 +212,16 @@ class BasicTest extends BaseTest
         $this->connection->executeStatement('TRUNCATE TABLE test_truncate_column');
 
         $event = $this->getEvent();
-        self::assertSame('CREATE TABLE test_truncate_column (id INTEGER(11), data VARCHAR(50))', $event->getQuery());
+        self::assertSame('CREATE TABLE test_truncate_column (id INTEGER(11), data VARCHAR(50))', $event->query);
         $event = $this->getEvent();
-        self::assertSame('BEGIN', $event->getQuery());
+        self::assertSame('BEGIN', $event->query);
         $event = $this->getEvent();
-        self::assertSame('TRUNCATE TABLE test_truncate_column', $event->getQuery());
+        self::assertSame('TRUNCATE TABLE test_truncate_column', $event->query);
     }
 
-    /**
-     * @test
-     */
-    public function shouldJsonSetPartialUpdateWithHoles(): void
+    public function testShouldJsonSetPartialUpdateWithHoles(): void
     {
-        if ($this->checkForVersion(5.7) || BinLogServerInfo::isMariaDb()) {
+        if ($this->checkForVersion(5.7) || $this->mySQLReplicationFactory?->getServerInfo()->isMariaDb()) {
             self::markTestIncomplete('Only for mysql 5.7 or higher');
         }
 
@@ -263,22 +242,16 @@ class BasicTest extends BaseTest
         $event = $this->getEvent();
 
         self::assertInstanceOf(UpdateRowsDTO::class, $event);
-        self::assertEquals(
-            $expected,
-            $event->getValues()[0]['before']['j']
-        );
+        self::assertEquals($expected, $event->values[0]['before']['j']);
         self::assertEquals(
             '{"age":22,"addr":{"code":100,"detail":{"ab":"970785C8"}},"name":"Alice"}',
-            $event->getValues()[0]['after']['j']
+            $event->values[0]['after']['j']
         );
     }
 
-    /**
-     * @test
-     */
-    public function shouldJsonRemovePartialUpdateWithHoles(): void
+    public function testShouldJsonRemovePartialUpdateWithHoles(): void
     {
-        if ($this->checkForVersion(5.7) || BinLogServerInfo::isMariaDb()) {
+        if ($this->checkForVersion(5.7) || $this->mySQLReplicationFactory?->getServerInfo()->isMariaDb()) {
             self::markTestIncomplete('Only for mysql 5.7 or higher');
         }
 
@@ -299,22 +272,16 @@ class BasicTest extends BaseTest
         $event = $this->getEvent();
 
         self::assertInstanceOf(UpdateRowsDTO::class, $event);
-        self::assertEquals(
-            $expected,
-            $event->getValues()[0]['before']['j']
-        );
+        self::assertEquals($expected, $event->values[0]['before']['j']);
         self::assertEquals(
             '{"age":22,"addr":{"code":100,"detail":{}},"name":"Alice"}',
-            $event->getValues()[0]['after']['j']
+            $event->values[0]['after']['j']
         );
     }
 
-    /**
-     * @test
-     */
-    public function shouldJsonReplacePartialUpdateWithHoles(): void
+    public function testShouldJsonReplacePartialUpdateWithHoles(): void
     {
-        if ($this->checkForVersion(5.7) || BinLogServerInfo::isMariaDb()) {
+        if ($this->checkForVersion(5.7) || $this->mySQLReplicationFactory?->getServerInfo()->isMariaDb()) {
             self::markTestIncomplete('Only for mysql 5.7 or higher');
         }
 
@@ -335,20 +302,14 @@ class BasicTest extends BaseTest
         $event = $this->getEvent();
 
         self::assertInstanceOf(UpdateRowsDTO::class, $event);
-        self::assertEquals(
-            $expected,
-            $event->getValues()[0]['before']['j']
-        );
+        self::assertEquals($expected, $event->values[0]['before']['j']);
         self::assertEquals(
             '{"age":22,"addr":{"code":100,"detail":{"ab":"9707"}},"name":"Alice"}',
-            $event->getValues()[0]['after']['j']
+            $event->values[0]['after']['j']
         );
     }
 
-    /**
-     * @test
-     */
-    public function shouldRoteLog(): void
+    public function testShouldRotateLog(): void
     {
         $this->connection->executeStatement('FLUSH LOGS');
 
@@ -356,14 +317,14 @@ class BasicTest extends BaseTest
 
         self::assertMatchesRegularExpression(
             '/^[a-z-]+\.[\d]+$/',
-            $this->getEvent()->getEventInfo()->getBinLogCurrent()->getBinFileName()
+            $this->getEvent()
+                ->getEventInfo()
+                ->binLogCurrent
+                ->getBinFileName()
         );
     }
 
-    /**
-     * @test
-     */
-    public function shouldUseProvidedEventDispatcher(): void
+    public function testShouldUseProvidedEventDispatcher(): void
     {
         $this->disconnect();
 
@@ -381,7 +342,7 @@ class BasicTest extends BaseTest
         /** @var QueryDTO $event */
         $event = $this->getEvent();
         self::assertInstanceOf(QueryDTO::class, $event);
-        self::assertEquals($createExpected, $event->getQuery());
+        self::assertEquals($createExpected, $event->query);
     }
 
     private function connectWithProvidedEventDispatcher(EventDispatcherInterface $eventDispatcher): void
@@ -393,13 +354,21 @@ class BasicTest extends BaseTest
             $eventDispatcher
         );
 
-        $this->connection = $this->mySQLReplicationFactory->getDbConnection();
+        $connection = $this->mySQLReplicationFactory->getDbConnection();
+        if ($connection === null) {
+            throw new RuntimeException('Connection not initialized');
+        }
 
+        $this->connection = $connection;
         $this->connection->executeStatement('SET SESSION time_zone = "UTC"');
         $this->connection->executeStatement('DROP DATABASE IF EXISTS ' . $this->database);
         $this->connection->executeStatement('CREATE DATABASE ' . $this->database);
         $this->connection->executeStatement('USE ' . $this->database);
         $this->connection->executeStatement('SET SESSION sql_mode = \'\';');
+
+        if ($this->checkForVersion(8.4) && $this->mySQLReplicationFactory->getServerInfo()->isGeneric()) {
+            self::assertInstanceOf(RotateDTO::class, $this->getEvent());
+        }
 
         self::assertInstanceOf(FormatDescriptionEventDTO::class, $this->getEvent());
         self::assertInstanceOf(QueryDTO::class, $this->getEvent());
